Refactoring done from plantplaces --> diaryqueen
