+++ src/main/java/com/diaryqueen/DiaryQueenApplication.java
@@ -1,13 +1,13 @@
-package com.plantplaces;
+package com.diaryqueen;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 
 @SpringBootApplication
-public class PlantplacesApplication {
+public class DiaryQueenApplication {
 
 	public static void main(String[] args) {
-		SpringApplication.run(PlantplacesApplication.class, args);
+		SpringApplication.run(DiaryQueenApplication.class, args);
 	}
 
 }
