+++ src/main/resources/templates/start.html
@@ -2,10 +2,10 @@
 <html xmlns:th="http://www.thymeleaf.org">
 <head>
 <meta charset="ISO-8859-1">
-<title>PlantPlaces.com</title>
+<title>diaryqueen.com</title>
 </head>
 <body>
-<h1>Welcome to PlantPlaces</h1>
+<h1>Welcome to diaryqueen</h1>
 <br/>
 <p th:text = "${specimenDTO}"/>
 </body>
