+++ src/test/java/com/diaryqueen/DiaryqueenApplicationTests.java
@@ -1,10 +1,10 @@
-package com.plantplaces;
+package com.diaryqueen;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class PlantplacesApplicationTests {
+class DiaryqueenApplicationTests {
 
 	@Test
 	void contextLoads() {
